refactor(scripts): migrate to CSV-based data loading and simplify tradeoff computation
Refactor evaluation scripts to use CSV format consistently and
simplify data processing pipeline.

- Rename p071_accuracy_visualize.py to p072_accuracy_visualize.py
- Remove JSON-based loading functions, use CSV files directly
- Simplify accuracy visualization to load from precomputed CSV
- Fix transpose issue in naive throughput aggregation
- Refactor tradeoff computation to use DataFrame operations
- Remove unused compute_tradeoff functions, save data directly
- Fix int() to float() conversion in throughput computation
- Add output directory cleanup in main workflow
- Remove debug print statements
- Clean up unused imports and type annotations

+++ scripts/p081_throughput_compute.py
@@ -162,7 +162,7 @@ def parse_runtime(df: pd.DataFrame, accessors: dict[str, Callable[[dict], list[d
         
         all_per_op.append(per_op)
 
-        total_time = int(per_op['time'].sum())
+        total_time = float(per_op['time'].sum())
         overall.append({
             'stage': stage,
             'dataset': dataset,
